requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
fastapi==0.86.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Pillow==9.3.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pytz==2022.6
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.20.4
typing_extensions==4.4.0
tzdata==2022.6
uvicorn==0.19.0
